requirements.txt
canonicalwebteam.flask-base==2.0.0
canonicalwebteam.templatefinder==1.0.0
canonicalwebteam.blog==6.4.4
canonicalwebteam.yaml-responses==1.2.0
Flask==2.3.3
Jinja2==3.1.4
djlint==1.34.1
canonicalwebteam.image-template==1.5.0
